# backend/swagger.json
{
    "openapi": "3.0.4",
    "info": {
        "title": "Loppis-sys API",
        "description": "API for a leaf market",
        "version": "1.0.0"
    },
    "tags": [
        {
            "name": "auth"
        },
        {
            "name": "sellers"
        }
    ],
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
        "/auth": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Return a JWT access token",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequestResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    },
                    "200": {
                        "description": "Access token created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sellers": {
            "post": {
                "tags": [
                    "sellers"
                ],
                "summary": "Create a customer",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    },
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ]
            },
            "get": {
                "tags": [
                    "sellers"
                ],
                "summary": "Returns a list of sellers.",
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    },
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "number"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ]
            }
        },
        "/sellers/{sellerId}": {
            "delete": {
                "tags": [
                    "sellers"
                ],
                "summary": "Delete a seller by id",
                "parameters": [
                    {
                        "name": "sellerId",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    },
                    "204": {}
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "sellers"
                ],
                "summary": "Update a seller by id",
                "parameters": [
                    {
                        "name": "sellerId",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    },
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "number"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ]
            },
            "get": {
                "tags": [
                    "sellers"
                ],
                "summary": "Returns a seller by id",
                "parameters": [
                    {
                        "name": "sellerId",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    },
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "name": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "number"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ]
            }
        },
        "/sellers/{sellerId}/products": {
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Add a product for a seller",
                "parameters": [
                    {
                        "name": "sellerId",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "category": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    },
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ]
            },
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Return products list from a seller",
                "parameters": [
                    {
                        "name": "sellerId",
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "403": {
                        "$ref": "#/components/responses/ForbiddenResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerErrorResponse"
                    },
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "category": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "number"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "responses": {
            "NotFoundResponse": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "InternalServerErrorResponse": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "BadRequestResponse": {
                "description": "Bad request",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "UnauthorizedResponse": {
                "description": "Not authorized. Authentication required",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "ForbiddenResponse": {
                "description": "Invalid or expired access token",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}